+++ tests/unit/__init__.py
@@ -2,4 +2,4 @@
 from tests.unit.base import TEST_DIR, BUILD_DIR, RESOURCES_DIR, BASE_CONFIG, ROOT_LOGGER, BZT_DIR
 from tests.unit.base import close_reader_file, local_paths_config
 from tests.unit.cases import BZTestCase, ExecutorTestCase
-from tests.unit.mocks import random_datapoint
+from tests.unit.mocks import random_datapoint, EngineEmul, BZMock
