+++ tests/unit/modules/test_consolidatingAggregator.py
@@ -3,8 +3,8 @@ from random import random, choice
 from apiritif import random_string
 from bzt.modules.aggregator import ConsolidatingAggregator, DataPoint, KPISet, AggregatorListener
 from bzt.utils import to_json
-from tests.unit import BZTestCase
-from tests.unit.mocks import r, MockReader, EngineEmul
+from tests.unit import BZTestCase, EngineEmul
+from tests.unit.mocks import r, MockReader
 
 
 def get_success_reader(offset=0):
