+++ tests/unit/modules/test_proxy2jmx.py
@@ -7,8 +7,7 @@ from bzt import TaurusConfigError
 from bzt.modules.proxy2jmx import Proxy2JMX, BZAProxy
 from bzt.modules.selenium import SeleniumExecutor
 from bzt.utils import is_windows, is_linux, get_full_path, RESOURCES_DIR
-from tests.unit import BZTestCase
-from tests.unit.mocks import EngineEmul
+from tests.unit import BZTestCase, EngineEmul
 
 
 class BZAProxyEmul(BZAProxy):
