+++ tests/unit/test_engine.py
@@ -7,8 +7,7 @@ import yaml
 from bzt import TaurusConfigError
 from bzt.engine import Configuration, EXEC
 from bzt.utils import BetterDict, is_windows, get_full_path, get_uniq_name, communicate
-from tests.unit import local_paths_config, RESOURCES_DIR, BZTestCase, ExecutorTestCase, TEST_DIR
-from tests.unit.mocks import EngineEmul
+from tests.unit import local_paths_config, RESOURCES_DIR, BZTestCase, ExecutorTestCase, TEST_DIR, EngineEmul
 
 
 class TestEngine(BZTestCase):
